Refactor to make way for new methods
