+++ src/lib/constants.js
@@ -1,0 +1,3 @@
+module.exports = {
+  VALID_EXTENSION_R: /\.(js|json)$/,
+};
