+++ src/lib/require-file.js
@@ -1,0 +1,9 @@
+// prettier-ignore
+const {
+  curry, compose,
+} = require('ramda');
+const { join2 } = require('../utils');
+
+const requireFile = curry((dir, file) => compose(require, join2(dir))(file));
+
+module.exports = requireFile;
